Support all Druid formats by using InputRows internally for default stacks.
It's now possible to make three different kinds of beam stacks with DruidBeams.

1. String/Array[Byte]/ByteBuffer based stacks. This lets you provide raw bytes,
   which will be converted to an InputRow using a Druid StringInputRowParser.
2. Map based stacks. This lets you provide a Map, which will be converted to
   an InputRow using a Druid MapInputRowParser.
3. Classic, custom-type stacks. This gives you full control over everything
   and is the kind of stack that has been with Tranquility since the beginning.

The first kind of stack is new. It's not currently a public API, because I'm
not really sure how best to expose it yet. It's used by Tranquility Kafka and
Tranquility Server.

This requires changing the Beam trait to prefer a new signature,

  def sendAll(messages: Seq[MessageType]): Seq[Future[SendResult]]

Because the old one (sendBatch) did not allow returning an error for some
messages but not others. This can now happen because the beam stack may
be taking Strings that may turn out to be unparseable. We want to return
errors for those unparseable Strings while still forwarding the rest.

Fixes #113, #119.